fix(js-utils): add self-contained @kinde/js-utils/expo entry
