+++ lib/expo.ts
@@ -1,6 +1,1 @@
 export { ExpoSecureStore } from "./sessionManager/stores/expoSecureStore.js";
-export { StorageKeys } from "./sessionManager/types.js";
-export type {
-  SessionManager,
-  TimeoutTokenData,
-} from "./sessionManager/types.js";
